Install python 2 and 3, 2 as default
Add python 3.x such that python 2.x is still the defualt for
a base python.exe use (such that webkit scripts that are not
yet both 2 and 3 supporting will keep working).

Make versions of the python executable available as python2
and python3 to allow user choice (pip is already available
as pip2 and pip3). This seems to match the general package
set ups for MacOS and Ubuntu.